pyproject.toml
[project]
name = "gws-meet2gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-firestore>=2.16.1",
    "google-cloud-storage>=2.18.2",
    "fastapi>=0.112.2",
    "uvicorn[standard]>=0.30.6",
    "supabase>=2.9.0",
    "google-genai>=1.28.0",
    "python-dotenv>=1.1.1",
]
